Update and rename rng.ts to rng.js
